requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
SQLAlchemy==2.0.36
alembic==1.14.0
passlib==1.7.4
pydantic==2.10.2
pydantic[email]
PyJWT==2.10.1
python_multipart==0.0.19
